Organize Main class and divide most of the content into the ShoePalaceBot class
