+++ src/main/java/io/forcesoftware/tasks/ShoePalaceTask.java
@@ -1,14 +1,13 @@
 package io.forcesoftware.tasks;
 
 import io.forcesoftware.Main;
+import io.forcesoftware.ShoePalaceBot;
 import io.forcesoftware.models.billing.Profile;
 import io.forcesoftware.models.task.TaskData;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class ShoePalaceTask implements Runnable {
-
-    private Thread thread;
+public class ShoePalaceTask extends Thread {
 
     private Logger logger;
 
@@ -17,13 +16,15 @@ public class ShoePalaceTask implements Runnable {
     private Profile profile;
 
     public ShoePalaceTask(TaskData taskData) {
+        super("ShoePalaceTask-" + taskData.getId());
+
         this.logger = LogManager.getLogger("io.forcesoftware.Task");
         this.taskData = taskData;
     }
 
     @Override
     public void run() {
-        for (Profile profile : Main.getInstance().getProfileLoader().getProfiles()) {
+        for (Profile profile : ShoePalaceBot.getInstance().getProfileLoader().getProfiles()) {
             if (profile.getAlias().equals(taskData.getProfileAlias())) {
                 this.profile = profile;
             }
@@ -35,12 +36,5 @@ public class ShoePalaceTask implements Runnable {
         }
 
         logger.info("Task started");
-    }
-
-    public void start() {
-        if (thread == null) {
-            thread = new Thread(this, "ShoePalaceTask-" + taskData.getId());
-            thread.start();
-        }
     }
 }
